# api/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a description.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Dispute Resolution Engine - v1
  version: "1.0"
paths:
...
  /dispute/{id}:
    get:
      consumes:
      - application/json
      description: Get a dispute
      parameters:
      - description: Dispute ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dispute Detail Endpoint
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get a dispute
      tags:
      - dispute
    patch:
      consumes:
      - application/json
      description: Update a dispute
      parameters:
      - description: Dispute ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dispute Patch Endpoint
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update a dispute
      tags:
      - dispute
...
swagger: "2.0"